V2019.1.0

+++ GameFramework/Resource/CheckResourcesCompleteCallback.cs
@@ -15,5 +15,5 @@ namespace GameFramework.Resource
     /// <param name="updateCount">要更新的资源数量。</param>
     /// <param name="updateTotalLength">要更新的资源总大小。</param>
     /// <param name="updateTotalZipLength">要更新的压缩包总大小。</param>
-    public delegate void CheckResourcesCompleteCallback(bool needUpdateResources, int removedCount, int updateCount, int updateTotalLength, int updateTotalZipLength);
+    public delegate void CheckResourcesCompleteCallback(bool needUpdateResources, int removedCount, int updateCount, long updateTotalLength, long updateTotalZipLength);
 }
